# contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Slack Off API
  version: 0.1.0
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
    TraceId:
      name: X-Trace-Id
      in: header
      required: false
      schema:
        type: string
  schemas:
...
    CheckInSession:
      type: object
      required: [id, status, startedAt, endedAt, durationSeconds, eligibleDurationSeconds, rewarded]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, completed, cancelled, invalidated]
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: [string, "null"]
          format: date-time
        durationSeconds:
          type: [integer, "null"]
        eligibleDurationSeconds:
          type: [integer, "null"]
        rewarded:
          type: boolean
    CheckInFinishResult:
      type: object
      required: [session, reward, fishTankOutcomes]
      properties:
        session:
          $ref: "#/components/schemas/CheckInSession"
        reward:
          type: object
          required: [score, drawProgress, rewarded]
          properties:
            score:
              type: integer
            drawProgress:
              type: integer
            drawChancesGranted:
              type: integer
            rewarded:
              type: boolean
            achievementsUnlocked:
              type: array
              items:
                $ref: "#/components/schemas/AchievementUnlock"
        fishTankOutcomes:
          type: array
          items:
            $ref: "#/components/schemas/FishTankResourceOutcome"
...
    AchievementUnlock:
      type: object
      required: [id, code, name, unlockedAt, rewards]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        unlockedAt:
          type: string
          format: date-time
        rewards:
          type: object
          required: [score, drawProgress, drawChances, cosmetic]
          properties:
            score:
              type: integer
            drawProgress:
              type: integer
            drawChances:
              type: integer
            cosmetic:
              type: [string, "null"]
...
    ActivityInteractionProgress:
      type: object
      additionalProperties: false
      properties:
        completedStepIds:
          type: array
          items:
            type: string
        timerSeconds:
          type: object
          additionalProperties:
            type: integer
        choiceAnswers:
          type: object
          additionalProperties:
            type: string
        miniGameResults:
          type: object
          additionalProperties:
            type: object
            properties:
              passed:
                type: boolean
              score:
                type: integer
        tapCounts:
          type: object
          additionalProperties:
            type: integer
        selectedOptions:
          type: object
          additionalProperties:
            type: string
        sortedItemIds:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        breathRounds:
          type: object
          additionalProperties:
            type: integer
        reactionResults:
          type: object
          additionalProperties:
            type: object
            required: [successCount, attempts]
            properties:
              successCount:
                type: integer
              attempts:
                type: integer
        journalEntries:
          type: object
          additionalProperties:
            type: object
            properties:
              text:
                type: string
              tagIds:
                type: array
                items:
                  type: string
...
    FishTankResourceType:
      type: string
      enum: [food, bubble, hatch_progress]
    FishTankResourceOutcome:
      type: object
      required: [resourceType, quantity, label, copy]
      properties:
        resourceType:
          $ref: "#/components/schemas/FishTankResourceType"
        quantity:
          type: integer
          minimum: 1
        label:
          type: string
        copy:
          type: string
...
paths:
...
  /v1/check-ins:
    post:
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/TraceId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                idempotencyKey:
                  type: string
                  minLength: 8
                  maxLength: 128
                interaction:
                  $ref: "#/components/schemas/ActivityInteractionProgress"
                theme:
                  type: string
                  enum: [office, restroom, daydream]
      responses:
        "200":
          description: Started check-in
          content:
            application/json:
              schema:
                type: object
                required: [data, error]
                properties:
                  data:
                    $ref: "#/components/schemas/CheckInSession"
                  error:
                    type: "null"
        "409":
          description: Active check-in already exists
  /v1/check-ins/{id}/finish:
    post:
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/TraceId"
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                interaction:
                  $ref: "#/components/schemas/ActivityInteractionProgress"
      responses:
        "200":
          description: Finished check-in with rewards
          content:
            application/json:
              schema:
                type: object
                required: [data, error]
                properties:
                  data:
                    $ref: "#/components/schemas/CheckInFinishResult"
                  error:
                    type: "null"
              examples:
                granted:
                  summary: Eligible check-in grants food
                  value:
                    data:
                      session:
                        id: "33333333-3333-4333-8333-333333333333"
                        status: completed
                        startedAt: "2026-07-16T10:00:00.000Z"
                        endedAt: "2026-07-16T10:05:00.000Z"
                        durationSeconds: 300
                        eligibleDurationSeconds: 300
                        rewarded: true
                      reward:
                        score: 5
                        drawProgress: 1
                        drawChancesGranted: 0
                        rewarded: true
                        achievementsUnlocked: []
                      fishTankOutcomes:
                        - resourceType: food
                          quantity: 1
                          label: "鱼食"
                          copy: "打卡完成，鱼食 +1。"
                    error: null
                unrewarded:
                  summary: Below-minimum check-in grants nothing
                  value:
                    data:
                      session:
                        id: "33333333-3333-4333-8333-333333333333"
                        status: completed
                        startedAt: "2026-07-16T10:00:00.000Z"
                        endedAt: "2026-07-16T10:00:10.000Z"
                        durationSeconds: 10
                        eligibleDurationSeconds: 0
                        rewarded: false
                      reward:
                        score: 0
                        drawProgress: 0
                        drawChancesGranted: 0
                        rewarded: false
                        achievementsUnlocked: []
                      fishTankOutcomes: []
                    error: null
                replay:
                  summary: Repeated finish returns original fish outcome
                  value:
                    data:
                      session:
                        id: "33333333-3333-4333-8333-333333333333"
                        status: completed
                        startedAt: "2026-07-16T10:00:00.000Z"
                        endedAt: "2026-07-16T10:05:00.000Z"
                        durationSeconds: 300
                        eligibleDurationSeconds: 300
                        rewarded: true
                      reward:
                        score: 0
                        drawProgress: 0
                        drawChancesGranted: 0
                        rewarded: true
                        achievementsUnlocked: []
                      fishTankOutcomes:
                        - resourceType: food
                          quantity: 1
                          label: "鱼食"
                          copy: "打卡完成，鱼食 +1。"
                    error: null
        "404":
          description: Check-in not found
  /v1/check-ins/active:
    get:
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Active check-in or null
          content:
            application/json:
              schema:
                type: object
                required: [data, error]
                properties:
                  data:
                    anyOf:
                      - $ref: "#/components/schemas/CheckInSession"
                      - type: "null"
                  error:
                    type: "null"
...